backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
python-multipart==0.0.20
pydantic==2.11.7
pydantic-settings==2.10.1
pdfplumber==0.11.7
python-docx==1.2.0
spacy==3.8.13
sentence-transformers==5.1.0
scikit-learn==1.7.1
numpy==2.3.2
pytest==8.4.1
httpx==0.28.1
